Initial commit - adds refactoring changes to enable in-app illuminate development
