+++ graylog2-server/src/main/java/org/graylog/plugins/pipelineprocessor/db/mongodb/MongoDbServicesModule.java
@@ -16,6 +16,9 @@
  */
 package org.graylog.plugins.pipelineprocessor.db.mongodb;
 
+import com.google.inject.multibindings.OptionalBinder;
+import org.graylog.plugins.pipelineprocessor.db.PaginatedPipelineService;
+import org.graylog.plugins.pipelineprocessor.db.PaginatedRuleService;
 import org.graylog.plugins.pipelineprocessor.db.PipelineService;
 import org.graylog.plugins.pipelineprocessor.db.PipelineStreamConnectionsService;
 import org.graylog.plugins.pipelineprocessor.db.RuleService;
@@ -24,8 +27,15 @@ import org.graylog2.plugin.PluginModule;
 public class MongoDbServicesModule extends PluginModule {
     @Override
     protected void configure() {
-        bind(PipelineService.class).to(MongoDbPipelineService.class);
-        bind(RuleService.class).to(MongoDbRuleService.class);
+        OptionalBinder.newOptionalBinder(binder(), PipelineService.class)
+                .setDefault().to(MongoDbPipelineService.class);
+        OptionalBinder.newOptionalBinder(binder(), RuleService.class)
+                .setDefault().to(MongoDbRuleService.class);
+        //bind(RuleService.class).to(MongoDbRuleService.class);
+        OptionalBinder.newOptionalBinder(binder(), PaginatedPipelineService.class)
+                .setDefault().to(PaginatedMongoDbPipelineService.class);
+        OptionalBinder.newOptionalBinder(binder(), PaginatedRuleService.class)
+                .setDefault().to(PaginatedMongoDbRuleService.class);
         bind(PipelineStreamConnectionsService.class).to(MongoDbPipelineStreamConnectionsService.class);
     }
 }
